+++ src/Pathing/Pathfinding.hpp
@@ -5,7 +5,7 @@
 namespace Pathing {
     class Pathfinding {
       public:
-        static float findDistance(Node &nodeA, Node &nodeB);
+        static int findDistance(Node &nodeA, Node &nodeB);
         static std::vector<Node*> findPath(Grid &nodeGrid, Node &startNode, Node &endNode);
         static bool containsNode(std::vector<Node*> &set, Node *node);
